requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
alembic==1.7.7
altair==4.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.10.1
atpublic==3.0.1
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.2
cycler==0.11.0
databricks-cli==0.16.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
docker==5.0.3
dpath==2.0.6
dulwich==0.20.40
dvc==2.10.2
dvc-render==0.0.5
dvclive==0.8.1
entrypoints==0.4
executing==0.8.3
Flask==2.0.3
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
mailchecker==4.1.17
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.24.0
multidict==6.0.2
nanotime==0.5.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.8.2
notebook==6.4.10
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.48
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pygit2==1.9.1
Pygments==2.11.2
pygtrie==2.4.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-benedict==0.25.1
python-dateutil==2.8.2
python-fsutil==0.6.1
python-slugify==6.1.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
requests==2.27.1
rich==12.4.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
scmrepo==0.0.19
semver==2.13.0
Send2Trash==1.8.0
shortuuid==1.0.9
shtab==1.5.4
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.7.0
tabulate==0.8.9
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
validators==0.18.2
voluptuous==0.13.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
xmltodict==0.13.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
